Feature/usuario (#11)
* feature(usuarios): implementadas as rotas de usuário e requisição de cadastro dos professores

* feature(foto|auth): configuradas as rotas de foto do usuário e criadas as autenticações de cargo.

* fix(foto): Solucionado problema com o salvamento das fotos e adicionadas validações de cargo nas rotas de usuário

* feature(email): básico da verificação por email feita.

* feature(email): adicionadas as entidades 'RequisicaoUsuario' e 'RequisicaoMudancaSenha' bem como suas rotas.

* fix(email): adicionado o motivo da recusa ao rejeitar requisição de cadastro de professor.

* fix: pequenas correções (incompleto)

* fix(MudançaSenha): adicionado o campo 'status' à requisição de mudança de senha.

* fix(MudançaSenha): Removida rota duplicada

* feature(templates): adicionados os templates de email.

* fix(repertorio): corrigida lógica de filtragem dos campos "likeDoUsuario" e "favoritadoPeloUsuario"

* feat(usuario): Adicionado rota para ver foto do usuario com Id

* fix(usuario): corrigido caminho da foto do user sendo retornado na resposta

* minor fixes

---------

+++ src/shared/controllers/UsuarioController.ts
@@ -1,11 +1,14 @@
+import { formatObraDoc } from '../../features/Repertorios/Helpers/FormatObraDoc';
 import { UsuarioService } from '../services/UsuarioService';
 import type {
   Controller,
   createUsuarioBody,
   professorCreateBody,
   RequestUserData,
+  updateSenhaBody,
   updateUsuarioBody,
 } from '../Types';
+import path from 'node:path';
 
 export const UsuarioController: Controller = {
   get: async (request, reply) => {
@@ -15,7 +18,7 @@ export const UsuarioController: Controller = {
     if (!response.success) {
       return reply
         .status(response.status as number)
-        .send({ message: response.message });
+        .send({ error: response.message });
     }
 
     return reply.status(200).send(response.data);
@@ -32,7 +35,7 @@ export const UsuarioController: Controller = {
         .send({ error: response.message });
     }
 
-    return reply.status(201).send({ message: response.message });
+    return reply.status(201).send({ id: response.data });
   },
 
   professorCreate: async (request, reply) => {
@@ -44,15 +47,37 @@ export const UsuarioController: Controller = {
   },
 
   update: async (request, reply) => {
-    const { nome, email, senha } = request.body as updateUsuarioBody;
+    const { nome, email } = request.body as updateUsuarioBody;
     const { id } = request.params as { id: string };
+    const { id: requisitor } = request.user as RequestUserData;
 
-    const response = await UsuarioService.update(id, { nome, email, senha });
+    if (requisitor !== id) {
+      return reply.status(403).send({
+        error: 'Não é possível editar informações de outro usuário.',
+      });
+    }
+
+    const response = await UsuarioService.update(id, { nome, email });
 
     if (!response.success) {
       return reply
         .status(response.status as number)
-        .send({ message: response.message });
+        .send({ error: response.message });
+    }
+
+    return reply.status(200).send({ message: response.message });
+  },
+
+  updateSenha: async (request, reply) => {
+    const { senha } = request.body as updateSenhaBody;
+    const { id } = request.params as { id: string };
+
+    const response = await UsuarioService.updateSenha(id, senha);
+
+    if (!response.success) {
+      return reply
+        .status(response.status as number)
+        .send({ error: response.message });
     }
 
     return reply.status(200).send({ message: response.message });
@@ -60,28 +85,55 @@ export const UsuarioController: Controller = {
 
   delete: async (request, reply) => {
     const { id } = request.params as { id: string };
+    const { id: requisitante, cargo } = request.user as RequestUserData;
+
+    if (requisitante !== id && cargo !== 'admin') {
+      return reply.status(403).send({
+        error: 'Não é possível editar a conta de outro usuário.',
+      });
+    }
 
     const response = await UsuarioService.delete(id);
 
     if (!response.success) {
       return reply
         .status(response.status as number)
-        .send({ message: response.message });
+        .send({ error: response.message });
     }
 
     return reply.status(200).send({ message: response.message });
   },
 
+  fotoGet: async (request, reply) => {
+    const { id } = request.params as { id: string };
+    const response = await UsuarioService.fotoGet(id);
+    if (!response.success) {
+      return reply
+        .status(response.status as number)
+        .send({ error: response.message });
+    }
+
+    const foto = response.data as string;
+    return reply.sendFile(foto);
+  },
+
   fotoCreate: async (request, reply) => {
     const { id } = request.params as { id: string };
+    const { id: requisitor } = request.user as RequestUserData;
     const files = await request.saveRequestFiles();
+
+    if (requisitor !== id) {
+      return reply.status(403).send({
+        error: 'Não é possível editar informações de outro usuário.',
+      });
+    }
 
     const response = await UsuarioService.fotoCreate(id, files[0]);
 
     if (!response.success) {
       return reply
         .status(response.status as number)
-        .send({ message: response.message });
+        .send({ error: response.message });
     }
 
     return reply.status(200).send({ message: response.message });
@@ -89,14 +141,21 @@ export const UsuarioController: Controller = {
 
   fotoUpdate: async (request, reply) => {
     const { id } = request.params as { id: string };
+    const { id: requisitor } = request.user as RequestUserData;
     const files = await request.saveRequestFiles();
+
+    if (requisitor !== id) {
+      return reply.status(403).send({
+        error: 'Não é possível editar informações de outro usuário.',
+      });
+    }
 
     const response = await UsuarioService.fotoUpdate(id, files[0]);
 
     if (!response.success) {
       return reply
         .status(response.status as number)
-        .send({ message: response.message });
+        .send({ error: response.message });
     }
 
     return reply.status(200).send({ message: response.message });
@@ -104,6 +163,13 @@ export const UsuarioController: Controller = {
 
   fotoDelete: async (request, reply) => {
     const { id } = request.params as { id: string };
+    const { id: requisitor } = request.user as RequestUserData;
+
+    if (requisitor !== id) {
+      return reply.status(403).send({
+        error: 'Não é possível editar informações de outro usuário.',
+      });
+    }
 
     const response = await UsuarioService.fotoDelete(id);
